+++ src/0-routes/__root.tsx
@@ -6,11 +6,11 @@ import {
 	Scripts,
 } from "@tanstack/react-router";
 import { TanStackRouterDevtoolsPanel } from "@tanstack/react-router-devtools";
-import Header from "../components/Header";
+import Header from "@/1-components/Header";
 
-import ConvexProvider from "../integrations/convex/provider";
-import TanStackQueryDevtools from "../integrations/tanstack-query/devtools";
-import appCss from "../styles.css?url";
+import ConvexProvider from "@/2-integrations/convex/provider";
+import TanStackQueryDevtools from "@/2-integrations/tanstack-query/devtools";
+import appCss from "@/styles.css?url";
 
 interface MyRouterContext {
 	queryClient: QueryClient;
