+++ src/0-routes/demo/form.address.tsx
@@ -1,6 +1,6 @@
 import { createFileRoute } from "@tanstack/react-router";
 
-import { useAppForm } from "@/hooks/demo.form";
+import { useAppForm } from "@/3-hooks/demo.form";
 
 export const Route = createFileRoute("/demo/form/address")({
 	component: AddressForm,
