+++ src/0-routes/demo/form.simple.tsx
@@ -1,7 +1,7 @@
 import { createFileRoute } from "@tanstack/react-router";
 import { z } from "zod";
 
-import { useAppForm } from "@/hooks/demo.form";
+import { useAppForm } from "@/3-hooks/demo.form";
 
 export const Route = createFileRoute("/demo/form/simple")({
 	component: SimpleForm,
