+++ src/0-routes/demo/start.ssr.data-only.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from "@tanstack/react-router";
-import { getPunkSongs } from "@/data/demo.punk-songs";
+import { getPunkSongs } from "@/5-data/demo.punk-songs";
 
 export const Route = createFileRoute("/demo/start/ssr/data-only")({
 	ssr: "data-only",
