+++ src/1-components/demo.FormComponents.tsx
@@ -1,6 +1,6 @@
 import { useStore } from "@tanstack/react-form";
 
-import { useFieldContext, useFormContext } from "@/hooks/demo.form-context";
+import { useFieldContext, useFormContext } from "@/3-hooks/demo.form-context";
 
 export function SubscribeButton({ label }: { label: string }) {
 	const form = useFormContext();
