--- src/routes/index.tsx
@@ -1,10 +1,0 @@
-import { createFileRoute } from "@tanstack/react-router";
-import { ComponentExample } from "@/components/component-example";
-
-export const Route = createFileRoute("/")({ component: App });
-
-function App() {
-return (
-  <ComponentExample />
-);
-}
